backend/requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
pymongo==4.11
requests==2.32.3
sentence_transformers
spacy
google-generativeai
numpy
torch
transformers